fixed warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@3497 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtui/InstanceDlg.cpp
@@ -203,7 +203,7 @@ void InstanceDlg::OnRadio( wxCommandEvent &event )
 
 vtContentManager *InstanceDlg::Current()
 {
-	if (m_iManager < m_contents.size())
+	if (m_iManager < (int) m_contents.size())
 		return m_contents[m_iManager];
 	return NULL;
 }
